+++ eras/conway/test-suite/test/Tests.hs
@@ -6,6 +6,7 @@ module Main where
 import Cardano.Ledger.Conway (Conway)
 import Data.Proxy (Proxy (..))
 import qualified Test.Cardano.Ledger.Babbage.TxInfo as Babbage (txInfoTests)
+import Test.Cardano.Ledger.Conway.GovSnapshot (govSnapshotProps)
 import qualified Test.Cardano.Ledger.Conway.Serialisation.CDDL as CDDL
 import qualified Test.Cardano.Ledger.Conway.Serialisation.Roundtrip as Roundtrip
 import qualified Test.Cardano.Ledger.Conway.TxInfo as Conway (txInfoTests)
@@ -22,4 +23,5 @@ defaultTests =
     , CDDL.tests 5
     , Babbage.txInfoTests (Proxy @Conway)
     , Conway.txInfoTests (Proxy @Conway)
+    , govSnapshotProps
     ]